Rename the design system's Pagination to Pager, after what it is
Two components were called Pagination and they are not two versions of one
thing:

  design-system  Prev/Next buttons over a `$bindable page`. Nothing touches the
                 URL, so the pages it steps through exist only while the
                 component is mounted. No consumer, in web or the extension.
  web            Real `<a href="?page=N">` in the SSR markup, window elision,
                 rel=prev/next. Four listings render it, and it is the only way
                 through any of them.

The shared name read as duplication, and the obvious next move on duplication is
to merge — which here means trading crawlable links for buttons a crawler cannot
press, on the four routes that depend on them. I wrote "reconciling them is its
own change" into #2040's description and was wrong: there is nothing to
reconcile, only a name that misdescribes one of them.

So the primitive is `Pager` now, and both files say in a comment what the other
one is for and why they stayed apart. Nothing imported the old name — the
adoption baseline recorded it at 0 consumers — so this renames a file, an
export, its test, its story and its DSDS docs entry, and touches no call site.

The `aria-label="Pagination"` inside it stays. That is what a screen reader
announces, and "Pager" is our word for the shape, not the user's for the control.

Kept, not deleted: four other primitives sit at 0 consumers and the baseline
records that as a normal state, so removing one of the five would be a policy
change wearing a cleanup's clothes.
