+++ design-system/scripts/adoption-baseline.json
@@ -15,7 +15,7 @@
   "LoadMore": 6,
   "NoticeDialog": 0,
   "NumberedGrid": 5,
-  "Pagination": 0,
+  "Pager": 0,
   "ProviderIcon": 8,
   "SectionLabel": 8,
   "SettingRow": 3,
